requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
BingImageCreator==0.1.5
certifi==2023.5.7
charset-normalizer==3.1.0
EdgeGPT==0.3.8
frozenlist==1.3.3
GoogleBard==1.0.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
prompt-toolkit==3.0.38
Pygments==2.15.1
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
rich==13.3.5
sniffio==1.3.0
termcolor==2.3.0
urllib3==2.0.2
wcwidth==0.2.6
websockets==11.0.3
yarl==1.9.2